@@ -1,143 +1,66 @@
 {
-    "name": "reaql-frontend",
-    "version": "0.0.1",
-    "author": "dan-exxx1st",
-    "license": "MIT",
-    "scripts": {
-        "dev": "next",
-        "dev-codegen": "npm run codegen && next",
-        "build": "npm run codegen && next build",
-        "start": "next start",
-        "storybook": "start-storybook -p 6006",
-        "build-storybook": "build-storybook",
-        "test": "jest",
-        "test:watch": "jest --watchAll",
-        "test:coverage": "npm run test --runInBand --coverage --collectCoverageFrom=src/**/*.{ts,tsx} --collectCoverageFrom=!src/**/*.stories.{ts,tsx} --collectCoverageFrom=!src/lib/**/* --collectCoverageFrom=!src/config/**/* --collectCoverageFrom=!src/lib --collectCoverageFrom=!src/**/*.{types,d}.{ts,tsx}",
-        "test:coverage:watch": "npm run test:coverage --watchAll",
-        "lint": "eslint \"src/**/*.{ts,tsx}\" --fix",
-        "pretify": "prettier --config=./.prettierrc --write ./src",
-        "codegen": "graphql-codegen"
-    },
-    "dependencies": {
-        "@apollo/client": "^3.1.4",
-        "@apollo/react-common": "3.1.4",
-        "@apollo/react-components": "^3.1.5",
-        "@babel/runtime": "^7.11.2",
-        "@svgr/webpack": "^5.4.0",
-        "apollo-cache": "^1.3.5",
-        "apollo-cache-inmemory": "^1.6.6",
-        "apollo-client": "^2.6.10",
-        "apollo-link": "1.2.14",
-        "apollo-link-http": "1.5.17",
-        "apollo-link-schema": "1.2.5",
-        "apollo-server-micro": "^2.14.2",
-        "apollo-utilities": "^1.3.3",
-        "cookie": "^0.4.1",
-        "date-fns": "^2.16.1",
-        "emoji-picker-react": "^3.2.4",
-        "enzyme": "^3.11.0",
-        "enzyme-adapter-react-16": "^1.15.3",
-        "final-form": "^4.20.1",
-        "graphql": "^14.6.0",
-        "jest": "^26.4.2",
-        "next": "latest",
-        "react": "^16.13.1",
-        "react-apollo": "^3.1.5",
-        "react-dom": "^16.13.1",
-        "react-final-form": "^6.5.2",
-        "styled-components": "^5.1.1"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.11.4",
-        "@babel/plugin-proposal-class-properties": "^7.12.1",
-        "@babel/plugin-transform-runtime": "^7.11.0",
-        "@graphql-codegen/cli": "^1.19.4",
-        "@graphql-codegen/plugin-helpers": "^1.15.1",
-        "@graphql-codegen/typescript": "^1.13.3",
-        "@graphql-codegen/typescript-apollo-client-helpers": "^1.1.1",
-        "@graphql-codegen/typescript-document-nodes": "^1.17.8",
-        "@graphql-codegen/typescript-operations": "^1.13.3",
-        "@graphql-codegen/typescript-react-apollo": "^2.2.1",
-        "@graphql-codegen/typescript-resolvers": "^1.15.1",
-        "@storybook/addon-actions": "^6.1.9",
-        "@storybook/addon-essentials": "^6.1.9",
-        "@storybook/addon-links": "^6.1.9",
-        "@storybook/react": "^6.1.9",
-        "@types/cookie": "^0.4.0",
-        "@types/enzyme": "^3.10.5",
-        "@types/enzyme-adapter-react-16": "^1.0.6",
-        "@types/jest": "^26.0.10",
-        "@types/react": "^16.9.34",
-        "@types/react-dom": "^16.9.7",
-        "@types/styled-components": "^5.1.2",
-        "@typescript-eslint/eslint-plugin": "^3.10.1",
-        "@typescript-eslint/parser": "^3.10.1",
-        "apollo": "^2.31.1",
-        "babel-loader": "^8.1.0",
-        "babel-plugin-styled-components": "^1.11.1",
-        "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
-        "enzyme-to-json": "^3.5.0",
-        "eslint": "^7.7.0",
-        "eslint-plugin-react": "^7.20.6",
-        "graphql-let": "0.x",
-        "husky": "^4.2.5",
-        "jest-styled-components": "^7.0.3",
-        "react-is": "^16.13.1",
-        "regenerator-runtime": "^0.13.7",
-        "ts-loader": "^8.0.3",
-        "typescript": "^3.8.3"
-    },
-    "jest": {
-        "roots": [
-            "<rootDir>/src"
-        ],
-        "collectCoverageFrom": [
-            "src/**/*.{js,jsx,ts,tsx}",
-            "!src/**/*.d.ts",
-            "!src/**/*.stories.{ts,tsx}"
-        ],
-        "setupFilesAfterEnv": [
-            "<rootDir>/src/setupTests.ts"
-        ],
-        "testMatch": [
-            "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
-            "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
-        ],
-        "transformIgnorePatterns": [
-            "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$",
-            "^.+\\.module\\.(css|sass|scss)$"
-        ],
-        "modulePaths": [
-            "<rootDir>/src"
-        ],
-        "moduleNameMapper": {
-            "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
-        },
-        "moduleFileExtensions": [
-            "web.js",
-            "js",
-            "web.ts",
-            "ts",
-            "web.tsx",
-            "tsx",
-            "json",
-            "web.jsx",
-            "jsx",
-            "node"
-        ],
-        "transform": {
-            "^.+\\.(js|jsx|ts|tsx)$": "<rootDir>/node_modules/babel-jest",
-            "^.+\\.css$": "<rootDir>/jest/cssTransform.js",
-            "^(?!.*\\.(js|jsx|ts|tsx|css|json)$)": "<rootDir>/jest/fileTransform.js"
-        },
-        "snapshotSerializers": [
-            "enzyme-to-json/serializer"
-        ]
-    },
-    "husky": {
-        "hooks": {
-            "pre-commit": "npm run lint",
-            "pre-push": "npm run test"
-        }
-    }
+  "name": "reaql-frontend",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@apollo/client": "^3.3.2",
+    "@testing-library/jest-dom": "^5.11.6",
+    "@testing-library/react": "^11.2.2",
+    "@testing-library/user-event": "^12.3.0",
+    "@types/jest": "^26.0.16",
+    "@types/node": "^12.19.8",
+    "@types/react": "^16.14.2",
+    "@types/react-dom": "^16.9.10",
+    "date-fns": "^2.16.1",
+    "graphql": "^15.4.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-scripts": "4.0.1",
+    "styled-components": "^5.2.1",
+    "typescript": "^4.1.2",
+    "web-vitals": "^0.2.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "storybook": "start-storybook -p 6006 -s public",
+    "build-storybook": "build-storybook -s public"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@storybook/addon-actions": "^6.1.9",
+    "@storybook/addon-essentials": "^6.1.9",
+    "@storybook/addon-links": "^6.1.9",
+    "@storybook/node-logger": "^6.1.9",
+    "@storybook/preset-create-react-app": "^3.1.5",
+    "@storybook/react": "^6.1.9",
+    "@types/date-fns": "^2.6.0",
+    "@types/enzyme": "^3.10.8",
+    "@types/enzyme-adapter-react-16": "^1.0.6",
+    "@types/styled-components": "^5.1.4",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
+    "enzyme-to-json": "^3.6.1",
+    "jest-styled-components": "^7.0.3",
+    "ts-loader": "^8.0.11"
+  }
 }
