@@ -1,26 +1,28 @@
 import React from 'react';
-import { AppProps } from 'next/app';
 import { ApolloProvider } from '@apollo/client';
 import { ThemeProvider } from 'styled-components';
 
-import { useApollo } from '../lib/graphql/apollo';
-import Themes from 'helpers/styled';
-import { StyledIndexWrapper } from 'helpers/styled/MainPageStyle';
 import { checkAuth } from 'helpers/authHelper';
 import AuthContext from 'helpers/contexts/authContext';
+import { StyledIndexWrapper } from 'helpers/styled/MainPageStyle';
+import { CreateApolloClient } from 'lib/graphql/apollo';
+import Theme from 'helpers/styled';
 
-export default function App({ Component, pageProps }: AppProps) {
-    const apolloClient: any = useApollo(pageProps.initialApolloState);
+import { HomePage } from './pages';
 
+const App = () => {
+    const apolloClient = CreateApolloClient();
     return (
         <ApolloProvider client={apolloClient}>
-            <ThemeProvider theme={Themes}>
+            <ThemeProvider theme={Theme}>
                 <StyledIndexWrapper>
                     <AuthContext.Provider value={checkAuth()}>
-                        <Component {...pageProps} />
+                        <HomePage />
                     </AuthContext.Provider>
                 </StyledIndexWrapper>
             </ThemeProvider>
         </ApolloProvider>
     );
-}
+};
+
+export default App;
